refactor: Simplify TokenStream by moving tree building into parser
This refactoring dramatically simplifies the lexing pipeline by moving
tree-building logic from the pipeline into the line-based parser where
it's actually needed.

Changes:
- Create tree_builder module in line-based parser with tree construction logic
- Add parse_from_flat_tokens() entry point that accepts flat tokens
- Remove ToLineTokensMapper and IndentationToTreeMapper from pipeline
- Remove adapters_linebased.rs module (no longer needed)
- Update all pipeline specs to produce flat token streams only
- Update executor and builder to use new flat API
- Update processor.rs to use tree_builder for debug output

Benefits:
- TokenStream stays simple (flat Vec<(Token, Range)> only)
- Pipeline responsibility: token transformations only (no structure)
- Complexity localized to line-based parser (only consumer that needs it)
- Reference parser simplified (direct flat consumption, no unroll needed)
- Better separation of concerns
- Net reduction of 1465 lines of code

Fixes #186
