+++ src/lex/parsing/linebased.rs
@@ -14,5 +14,6 @@
 
 pub mod declarative_grammar;
 pub mod engine;
+pub mod tree_builder;
 
-pub use engine::parse_experimental_v2;
+pub use engine::{parse_experimental_v2, parse_from_flat_tokens};
