+++ src/lex/pipeline/mappers.rs
@@ -4,13 +4,9 @@
 //! that perform specific transformations on TokenStreams.
 
 pub mod blank_lines;
-pub mod indentation_to_tree;
 pub mod normalize_whitespace;
 pub mod semantic_indentation;
-pub mod to_line_tokens;
 
 pub use blank_lines::BlankLinesMapper;
-pub use indentation_to_tree::IndentationToTreeMapper;
 pub use normalize_whitespace::NormalizeWhitespaceMapper;
 pub use semantic_indentation::SemanticIndentationMapper;
-pub use to_line_tokens::ToLineTokensMapper;
